requirements.txt
asttokens==3.0.1
certifi==2026.4.22
charset-normalizer==3.4.7
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
idna==3.14
ipykernel==7.2.0
ipython==9.13.0
ipython_pygments_lexers==1.1.1
jedi==0.20.0
jupyter_client==8.8.0
jupyter_core==5.9.1
matplotlib-inline==0.2.2
nest-asyncio==1.6.0
numpy==2.4.4
packaging==26.2
pandas==3.0.2
parso==0.8.7
pexpect==4.9.0
platformdirs==4.9.6
prompt_toolkit==3.0.52
psutil==7.2.2
psycopg2-binary==2.9.12
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyzmq==27.1.0
requests==2.33.1
six==1.17.0
stack-data==0.6.3
tornado==6.5.5
traitlets==5.15.0
urllib3==2.7.0
wcwidth==0.7.0
